Move model type helpers into /internals
