+++ __tests-pacts__/serlo.org-database-layer/alias.ts
@@ -22,7 +22,7 @@
 import { Matchers } from '@pact-foundation/pact'
 
 import { addMessageInteraction } from '../__utils__'
-import { Payload } from '~/model'
+import { Payload } from '~/internals/model/types'
 import { Instance } from '~/types'
 
 test('AliasQuery', async () => {
