+++ api/api.serlo.org.api.md
@@ -8,7 +8,7 @@ import type { Context } from 'internals/graphql/context';
 import type { GraphQLResolveInfo } from 'graphql';
 import type { GraphQLScalarType } from 'graphql';
 import type { GraphQLScalarTypeConfig } from 'graphql';
-import type { ModelOf } from '~/model/types';
+import type { ModelOf } from '~/internals/model/types';
 
 // @public (undocumented)
 export type AbstractEntity = {
@@ -3546,6 +3546,72 @@ export type VideoThreadsArgs = {
     trashed?: Maybe<Scalars['Boolean']>;
 };
 
+
+// Warnings were encountered during analysis:
+//
+// __tests__/__utils__/handlers.ts:26:10 - (TS2305) Module '"~/model"' has no exported member 'Payload'.
+// src/internals/graphql/schema.ts:27:10 - (TS2305) Module '"~/model"' has no exported member 'Typename'.
+// src/schema/thread/resolvers.ts:47:8 - (TS7006) Parameter 'thread' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:50:15 - (TS7006) Parameter 'thread' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:53:11 - (TS7006) Parameter 'thread' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:56:14 - (TS7006) Parameter 'thread' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:59:18 - (TS7006) Parameter 'thread' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:59:26 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:59:35 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:68:14 - (TS7006) Parameter 'thread' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:68:22 - (TS7006) Parameter 'cursorPayload' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:70:45 - (TS7006) Parameter 'a' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:70:48 - (TS7006) Parameter 'b' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:73:18 - (TS2571) Object is of type 'unknown'.
+// src/schema/thread/resolvers.ts:80:15 - (TS7006) Parameter 'comment' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:83:18 - (TS7006) Parameter 'comment' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:83:27 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/thread/resolvers.ts:83:36 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
+// src/schema/uuid/abstract-navigation-child/resolvers.ts:34:10 - (TS7006) Parameter 'navigationChild' implicitly has an 'any' type.
+// src/schema/uuid/abstract-navigation-child/resolvers.ts:34:27 - (TS7006) Parameter 'cursorPayload' implicitly has an 'any' type.
+// src/schema/uuid/abstract-navigation-child/resolvers.ts:35:56 - (TS7006) Parameter 'node' implicitly has an 'any' type.
+// src/schema/uuid/abstract-navigation-child/resolvers.ts:35:62 - (TS7006) Parameter 'index' implicitly has an 'any' type.
+// src/schema/uuid/course-page/resolvers.ts:40:18 - (TS7006) Parameter 'coursePage' implicitly has an 'any' type.
+// src/schema/uuid/course-page/resolvers.ts:40:30 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/uuid/course-page/resolvers.ts:40:39 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
+// src/schema/uuid/course/resolvers.ts:43:17 - (TS7006) Parameter 'course' implicitly has an 'any' type.
+// src/schema/uuid/course/resolvers.ts:43:27 - (TS7031) Binding element 'trashed' implicitly has an 'any' type.
+// src/schema/uuid/course/resolvers.ts:43:36 - (TS7031) Binding element 'hasCurrentRevision' implicitly has an 'any' type.
+// src/schema/uuid/course/resolvers.ts:43:60 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
+// src/schema/uuid/course/resolvers.ts:54:38 - (TS2571) Object is of type 'unknown'.
+// src/schema/uuid/course/resolvers.ts:57:19 - (TS2571) Object is of type 'unknown'.
+// src/schema/uuid/exercise-group/resolvers.ts:43:21 - (TS7006) Parameter 'exerciseGroup' implicitly has an 'any' type.
+// src/schema/uuid/exercise-group/resolvers.ts:43:36 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/uuid/exercise-group/resolvers.ts:43:45 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
+// src/schema/uuid/grouped-exercise/resolvers.ts:42:25 - (TS7006) Parameter 'groupedExercise' implicitly has an 'any' type.
+// src/schema/uuid/grouped-exercise/resolvers.ts:42:42 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/uuid/grouped-exercise/resolvers.ts:42:51 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
+// src/schema/uuid/page/resolvers.ts:33:16 - (TS7006) Parameter 'page' implicitly has an 'any' type.
+// src/schema/uuid/page/resolvers.ts:33:22 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/uuid/page/resolvers.ts:33:31 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
+// src/schema/uuid/solution/resolvers.ts:38:20 - (TS7006) Parameter 'solution' implicitly has an 'any' type.
+// src/schema/uuid/solution/resolvers.ts:38:30 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/uuid/solution/resolvers.ts:38:39 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
+// src/schema/uuid/taxonomy-term/resolvers.ts:35:18 - (TS7006) Parameter 'taxonomyTerm' implicitly has an 'any' type.
+// src/schema/uuid/taxonomy-term/resolvers.ts:35:32 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/uuid/taxonomy-term/resolvers.ts:35:41 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
+// src/schema/uuid/taxonomy-term/resolvers.ts:41:20 - (TS7006) Parameter 'taxonomyTerm' implicitly has an 'any' type.
+// src/schema/uuid/taxonomy-term/resolvers.ts:41:34 - (TS7006) Parameter 'cursorPayload' implicitly has an 'any' type.
+// src/schema/uuid/taxonomy-term/resolvers.ts:41:51 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
+// src/schema/uuid/taxonomy-term/resolvers.ts:43:39 - (TS7006) Parameter 'id' implicitly has an 'any' type.
+// src/schema/uuid/taxonomy-term/resolvers.ts:51:18 - (TS2571) Object is of type 'unknown'.
+// src/schema/uuid/taxonomy-term/resolvers.ts:55:22 - (TS7006) Parameter 'taxonomyTerm' implicitly has an 'any' type.
+// src/schema/uuid/taxonomy-term/resolvers.ts:55:36 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/uuid/taxonomy-term/resolvers.ts:55:43 - (TS7006) Parameter 'context' implicitly has an 'any' type.
+// src/schema/uuid/user/resolvers.ts:71:24 - (TS7006) Parameter 'user' implicitly has an 'any' type.
+// src/schema/uuid/user/resolvers.ts:71:30 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/uuid/user/resolvers.ts:71:39 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
+// src/schema/uuid/user/resolvers.ts:76:23 - (TS7006) Parameter 'user' implicitly has an 'any' type.
+// src/schema/uuid/user/resolvers.ts:76:29 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/uuid/user/resolvers.ts:76:36 - (TS7006) Parameter 'context' implicitly has an 'any' type.
+// src/schema/uuid/user/resolvers.ts:81:26 - (TS7006) Parameter 'user' implicitly has an 'any' type.
+// src/schema/uuid/user/resolvers.ts:81:32 - (TS7006) Parameter '_args' implicitly has an 'any' type.
+// src/schema/uuid/user/resolvers.ts:81:41 - (TS7031) Binding element 'dataSources' implicitly has an 'any' type.
 
 // (No @packageDocumentation comment for this package)
